+++ src/main/java/com/falsepattern/zigbrains/zon/ZonLanguage.java
@@ -1,0 +1,10 @@
+package com.falsepattern.zigbrains.zon;
+
+import com.intellij.lang.Language;
+
+public class ZonLanguage extends Language {
+    public static final ZonLanguage INSTANCE = new ZonLanguage();
+    private ZonLanguage() {
+        super("Zon");
+    }
+}