+++ src/main/java/com/falsepattern/zigbrains/zon/lexer/ZonLexerAdapter.java
@@ -1,0 +1,9 @@
+package com.falsepattern.zigbrains.zon.lexer;
+
+import com.intellij.lexer.FlexAdapter;
+
+public class ZonLexerAdapter extends FlexAdapter {
+    public ZonLexerAdapter() {
+        super(new ZonFlexLexer(null));
+    }
+}