+++ src/main/java/com/falsepattern/zigbrains/zon/psi/ZonElementType.java
@@ -1,0 +1,12 @@
+package com.falsepattern.zigbrains.zon.psi;
+
+import com.falsepattern.zigbrains.zon.ZonLanguage;
+import com.intellij.psi.tree.IElementType;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+
+public class ZonElementType extends IElementType {
+    public ZonElementType(@NonNls @NotNull String debugName) {
+        super(debugName, ZonLanguage.INSTANCE);
+    }
+}